Update warcit.py
Adapted method make_index_revisit: When creating the revisit record the information of the WARC Date of the index/default file is passed on so that the revisit record has the same WARC-Date as the index file. This is helpful in Playback. The actual creation date is added is WARC-Creation-Date.